# backend/api/swagger.yaml
openapi: "3.0.3"
info:
  title: 算力撮合交易平台 API
  description: |
    ## 概述
    算力现货挂牌撮合交易平台 v1 API。包含账户认证、算力撮合、支付分账、居间金融、运营管理、区块链存证六大模块。

    **Base URL**: `http://localhost:8080/api/v1`

    ## 鉴权
    除公开接口外，所有接口需 `Authorization: Bearer <jwt_token>`

    ## 金额单位
    所有金额单位为**分（fen）**，即 `100 = ¥1.00`

    ## 通用响应格式
    ```json
    { "code": 0, "message": "success", "data": {}, "request_id": "xxx" }
    ```
  version: "1.0.0"
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Response:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: "success" }
        data: { type: object }
        request_id: { type: string }
...
    Order:
      type: object
      properties:
        id: { type: integer }
        order_no: { type: string, example: "ORD20260713143000123abc" }
        buyer_id: { type: integer }
        product_id: { type: integer }
        quantity: { type: integer }
        duration: { type: integer, description: "租期(小时)" }
        unit_price: { type: integer, description: "单价(分)" }
        total_amount: { type: integer, description: "总金额(分)" }
        platform_fee: { type: integer, description: "平台佣金(分)" }
        status: { type: string, enum: [pending_payment, paid, provisioning, active, completed, cancelled, refunding, refunded, frozen] }
        payment_expires_at: { type: string, format: date-time }
        lease_start_at: { type: string, format: date-time }
        lease_end_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    BuyerOrder:
      allOf:
        - { $ref: '#/components/schemas/Order' }
        - type: object
          properties:
            gpu_model: { type: string, description: "商品 GPU 型号；不适用时为空串" }
            product_type: { type: string, enum: [card_rental, outright, center, colocation] }
            pricing_mode: { type: string, enum: [hourly, daily, weekly, monthly, perpetual] }
            self_operated: { type: boolean }
            supplier_name: { type: string, description: "自营商品为平台自营；非自营取已认证企业名称，无资料时为空串" }
    BuyerOrderPage:
      type: object
      properties:
        list:
          type: array
          items: { $ref: '#/components/schemas/BuyerOrder' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    BuyerOrderDetail:
      type: object
      required: [order, product, supplier, delivery, actions]
      properties:
        order:
          type: object
          properties:
            order_no: { type: string, example: "ORD20260713143000123abc" }
            status: { type: string, enum: [pending_payment, paid, provisioning, active, completed, cancelled, refunding, refunded, frozen] }
            quantity: { type: integer }
            duration: { type: integer, description: "计费周期数，单位由 product.pricing_mode 决定" }
            unit_price: { type: integer, description: "单价（分）" }
            total_amount: { type: integer, description: "总金额（分）" }
            platform_fee: { type: integer, description: "平台佣金（分）" }
            payment_expires_at: { type: string, format: date-time, nullable: true }
            lease_start_at: { type: string, format: date-time, nullable: true }
            lease_end_at: { type: string, format: date-time, nullable: true }
            compliance_agreed: { type: boolean }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
        product:
          type: object
          description: "当前关联商品资料；现有数据模型没有不可变的下单快照"
          properties:
            id: { type: integer }
            product_type: { type: string, enum: [card_rental, outright, center, colocation] }
            gpu_model: { type: string }
            card_count: { type: integer }
            machine_count: { type: integer, nullable: true }
            total_pflops_approx: { type: string, nullable: true }
            power_capacity_kw: { type: integer, nullable: true }
            rack_count: { type: integer, nullable: true }
            cpu_spec: { type: string }
            memory_spec: { type: string }
            storage_spec: { type: string }
            bandwidth_spec: { type: string }
            delivery_mode: { type: string }
            pricing_mode: { type: string, enum: [hourly, daily, weekly, monthly, perpetual] }
            region: { type: string }
            self_operated: { type: boolean }
        supplier:
          type: object
          properties:
            name: { type: string }
            self_operated: { type: boolean }
            credit:
              type: object
              nullable: true
              properties:
                fulfill_rate: { type: number }
                sla_rate: { type: number }
                violation_count: { type: integer }
                total_orders: { type: integer }
                updated_at: { type: string, format: date-time }
        delivery:
          type: object
          nullable: true
          description: "仅返回状态摘要；访问凭证继续使用独立接口"
          properties:
            access_status: { type: string, enum: [none, generated, delivered, revoked] }
            access_expires_at: { type: string, format: date-time, nullable: true }
            revoked_at: { type: string, format: date-time, nullable: true }
            confirmed_by_buyer: { type: boolean }
            buyer_confirmed_at: { type: string, format: date-time, nullable: true }
            created_at: { type: string, format: date-time }
        actions:
          type: object
          properties:
            can_confirm: { type: boolean }
            can_renew: { type: boolean }
            can_refund: { type: boolean }
            can_view_credential: { type: boolean }
...
security:
  - BearerAuth: []
paths:
...
  /supplier-applications:
    get:
      tags: [算力撮合 Compute]
      summary: 我的供给方入驻申请
    post:
      tags: [算力撮合 Compute]
      summary: 提交供给方入驻申请
      description: "需先完成个人或企业 KYC；审核通过后自动授予 supplier 角色。"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [company_name, credit_code, representative, representative_id_number, business_license, contact_method, bank_name, account_name, account_number, facility_address, has_idc_license, power_description, cooling_description]
              properties:
                company_name: { type: string }
                credit_code: { type: string, minLength: 18, maxLength: 18 }
                representative: { type: string }
                representative_id_number: { type: string }
                business_license: { type: string, format: binary, description: "PDF/JPG/PNG，最大 5MB；文件内容写入 MySQL" }
                contact_method: { type: string }
                bank_name: { type: string }
                account_name: { type: string }
                account_number: { type: string }
                facility_address: { type: string }
                has_idc_license: { type: boolean }
                power_description: { type: string }
                cooling_description: { type: string }

  # ==================== 算力撮合 ====================
  /products:
    get:
      tags: [算力撮合 Compute]
      summary: 算力商品列表
      security: []
      parameters:
        - { name: gpu_model, in: query, schema: { type: string } }
        - { name: region, in: query, schema: { type: string } }
        - { name: pricing_mode, in: query, schema: { type: string, enum: [hourly, weekly, monthly] } }
        - { name: price_min, in: query, schema: { type: integer } }
        - { name: price_max, in: query, schema: { type: integer } }
        - { name: sort, in: query, schema: { type: string, enum: [price_asc, price_desc, created_at_desc] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: "商品列表" }
    post:
      tags: [算力撮合 Compute]
      summary: 上架算力商品（供给方）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [gpu_model, card_count, delivery_mode, pricing_mode, unit_price, stock, region, compliance_agreed]
              properties:
                gpu_model: { type: string }
                card_count: { type: integer }
                cpu_spec: { type: string }
                memory_spec: { type: string }
                storage_spec: { type: string }
                bandwidth_spec: { type: string }
                delivery_mode: { type: string }
                pricing_mode: { type: string }
                unit_price: { type: integer, description: "单价(分)/卡·时" }
                available_hours: { type: string }
                stock: { type: integer }
                min_order: { type: integer }
                min_duration: { type: integer }
                region: { type: string }
                compliance_agreed: { type: boolean }
  /products/{id}:
    get:
      tags: [算力撮合 Compute]
      summary: 商品详情（含信用分）
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: "OK" }
  /orders:
    post:
      tags: [算力撮合 Compute]
      summary: 下单
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [product_id, quantity, duration, compliance_agreed]
              properties:
                product_id: { type: integer }
                quantity: { type: integer }
                duration: { type: integer, description: "租期(小时)" }
                compliance_agreed: { type: boolean }
      responses:
        '200': { description: "订单创建成功" }
    get:
      tags: [算力撮合 Compute]
      summary: 我的订单（买家）
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending_payment, paid, provisioning, active, completed, cancelled, refunding, refunded, frozen] } }
        - { name: order_no, in: query, schema: { type: string }, description: "订单号包含搜索，可带开头的 #" }
        - { name: page, in: query, schema: { type: integer, minimum: 1, maximum: 1000000 }, description: "默认 1" }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100 }, description: "默认 20，最大 100" }
      responses:
        '200':
          description: "业务结果；仍需检查响应 code 是否为 0"
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BuyerOrderPage' }
  /orders/{order_no}:
    get:
      tags: [算力撮合 Compute]
      summary: 买家订单详情
      description: "仅返回当前 JWT 用户自己的订单；不存在或非本人均为 code=40400。商品/供给方是当前关联资料，不是下单快照。"
      parameters:
        - { name: order_no, in: path, required: true, schema: { type: string, pattern: '^(ORD|REN)[A-Za-z0-9-]{6,29}$', maxLength: 32 }, description: "订单号，不接受自增 ID" }
      responses:
        '200':
          description: "业务结果；仍需检查响应 code 是否为 0"
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Response' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BuyerOrderDetail' }
...
  /orders/{id}/confirm:
    post:
      tags: [算力撮合 Compute]
      summary: 确认签收（买家）
      description: "仅订单本人可确认 provisioning 且已生成访问凭证的订单。请传订单号，不传请求体。"
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: "订单号 order_no" }
      responses:
        '200': { description: "成功 code=0；不可签收/重复签收 code=40900；非本人 code=40300" }
  /orders/{id}/renew:
    post:
      tags: [算力撮合 Compute]
      summary: 续费
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [duration]
              properties:
                duration: { type: integer }
...
  /admin/audits/qualifications:
    get:
      tags: [运营管理 Admin]
      summary: 资质审核列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected, expired, all]
            default: pending
  /admin/audits/qualifications/{id}/document:
    get:
      tags: [运营管理 Admin]
      summary: 下载供给方申请附件
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
...